// src/lib/api/generated/model/emailTemplatesListParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

export type EmailTemplatesListParams = {
search?: string;
status?: string;
template_type?: string;
};